requirements.txt
Flask>=3.0.0
flask-limiter
flask-wtf
flask-talisman
qrcode>=7.4.2
pillow>=10.4.0
python-dotenv>=1.0.1
moviepy>=2.0.0
imageio>=2.35.1
imageio-ffmpeg>=0.5.1
numpy>=1.26.4
pywhatkit>=5.4
python-dotenv==1.0.1
pyautogui==0.9.54
pyperclip==1.9.0
python-docx>=1.1.2
gunicorn
